Updating postgresql to 9.6.3, adding systemd unit service.
Summary:
I've upgraded postgresql to the latest version. I've also created the systemd unit service file to start/stop/reload postgres using systemd. It's started by a script which checks to see if the /var/db/pgsql/data folder is empty, if it is, it calls initdb to initialize the database and get your default config files setup.

I started with the service file defined in [[ https://postgrespro.com/docs/postgrespro/9.6/server-start | postgres' docs ]]. I updated the ExecStart to call out to `postgresql.sh`, since the service was setup as `Type=notify`, I had to add a `NotifyAccess=all` to that it would allow any pid to send the notification to systemd. This seemed to work ok for starting, but reloading/stopping seemed to be sending the kill signals to the wrong pid. I attempted to fix that by defining a `PIDFile` and pointing it to the pid file `/var/db/pgsql/data/postmaster.pid` that postgres uses, but it didn't help. I ended up changing this to be a forking service and use pg_ctl to control everything. That seemed to do the trick, Start/Stop/Reload all work without any errors. If you do attempt to mess with getting the `Type=notify` working, be sure you compile postgres with the `--with-systemd` option.

All of this is to say that the unit file is going against the sample service file and method. That stems from the use of the script to start postgres so we can determine if we should run initdb first, instead of starting it directly, like the sample service file.

I believe this fixes T648

Test Plan:
I've tested this by doing the following steps.

1. make the package
2. install the package (before doing this, I'd ensure the logs and data folders were removed by running `sudo rm -rf /var/db/pgsql /var/log/pgsql`, but that's only because I was constantly removing/installing the package)
3. start postgres by running `sudo systemctl start postgresql`
4. ensuring it has started by checking `sudo systemctl status postgresql -l`
5. installing pgadmin3 and connecting to localhost:5432
6. I was also able to run a `mix ecto.create` in a phoenix project and the database was created properly in postgres
7. stopping/reloading postgres with `sudo systemctl [stop|reload] postgresql`

Reviewers: #triage_team, ikey

Reviewed By: #triage_team, ikey

Subscribers: ikey

Maniphest Tasks: T648

Differential Revision: https://dev.solus-project.com/D124